* instead of a single SSHException, now 2 subclasses TransportException and UserAuthException used by pertinent layers * make packet sequence numbers longs, and wrap around to 0 after 2^32 packets * started thoroughly documenting userauth.* * abstract base class for services * made Buffer put*() calls support method chaining * moved classes around (again!), method naming, etc

+++ src/main/java/org/apache/commons/net/ssh/mac/BaseMAC.java
@@ -103,7 +103,7 @@ public class BaseMAC implements MAC
         mac.update(foo, 0, foo.length);
     }
     
-    public void update(int i)
+    public void update(long i)
     {
         tmp[0] = (byte) (i >>> 24);
         tmp[1] = (byte) (i >>> 16);
